Added File Structure for DB
