+++ bin/seed.js
@@ -1,0 +1,12 @@
+const { db } = require('../server/db');
+
+const init = async () => {
+  try {
+    await db.sync({ force: true });
+    console.log('connected');
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+init();
