+++ package.json
@@ -8,6 +8,7 @@
     "test:dev": "npm run test -- --watch",
     "build": "webpack",
     "build:dev": "webpack --watch --mode=development",
+    "seed": "node ./bin/seed.js",
     "start": "node ./server/server.js",
     "start:dev": "npm run build:dev & nodemon ./server/server.js"
   },
