+++ server/db/hooks.js
@@ -1,0 +1,10 @@
+//Import models here from relationships.js
+
+/*
+
+Hooks that require more than one model go here. We can divide them into regions
+using //#region  if needed
+
+*/
+
+//Export models here and into index.js
