+++ server/db/relationships.js
@@ -1,0 +1,7 @@
+//Import all models from models folder here
+
+/*
+Relationships Here
+*/
+
+//After relationships are created, export them here into hooks.js
